requirements.txt
atlassian-python-api==3.41.16
requests>=2.25.1
flask==2.2.3
